added cookies test and renamed homepage file
