+++ src/Dfe.ManageSchoolImprovement.CypressTests/cypress/e2e/homePage.cy.ts
@@ -1,14 +1,14 @@
 import { Logger } from "cypress/common/logger";
-import riseHomePage from "cypress/pages/riseHomePage";
+import riseHomePage from "cypress/pages/homePage";
 import path from "path";
 
 
-describe("User navigates to the rise landing page", () => {
+describe("User navigates to the MSI landing page", () => {
     beforeEach(() => {
         cy.login();   
     });
 
-    it("Should be able to see Add a school option and school list", () => {
+    it("should have 'Add a school' button, school list and filter section", () => {
        riseHomePage
          .hasAddSchool()
          .hasProjectCount()
@@ -16,6 +16,12 @@ describe("User navigates to the rise landing page", () => {
          
          cy.executeAccessibilityTests()
     });
+
+    it("should have cookies banner and user should navigate to the cookies page", () => {
+       riseHomePage
+         .hasCookiesBanner()
+         .viewCookiesPage()
+    });    
 
     it("Should show the success notification banner when a filter is applied", () => {
         riseHomePage
